extracting duplicated Retriever in tests (#3382)
